move component into gpui
